requirements.txt
streamlit>=1.32.0
scikit-learn>=1.4.0
numpy>=1.26.0
matplotlib>=3.8.0
Pillow>=10.2.0